move phases to evaluation package
