+++ modules/expression-compiler/src/main/scala-2/scala/tools/nsc/ExpressionCompilerBridge.scala
@@ -4,6 +4,7 @@ import java.nio.file.Path
 import java.util.function.Consumer
 import java.{util => ju}
 import scala.jdk.CollectionConverters._
+import scala.tools.nsc.evaluation.ExpressionGlobal
 import scala.tools.nsc.reporters.StoreReporter
 import scala.util.control.NonFatal
