+++ modules/expression-compiler/src/main/scala-2/scala/tools/nsc/evaluation/ExpressionGlobal.scala
@@ -1,6 +1,8 @@
-package scala.tools.nsc
+package scala.tools.nsc.evaluation
 
+import scala.tools.nsc.Global
 import scala.tools.nsc.reporters.Reporter
+import scala.tools.nsc.Settings
 
 private[nsc] class ExpressionGlobal(
     settings: Settings,
