+++ modules/expression-compiler/src/main/scala-2/scala/tools/nsc/evaluation/GenerateExpression.scala
@@ -1,4 +1,4 @@
-package scala.tools.nsc
+package scala.tools.nsc.evaluation
 
 import scala.collection.mutable
 import scala.tools.nsc.transform.{Transform, TypingTransformers}
